+++ src/migrations/Install.php
@@ -20,8 +20,7 @@ class Install extends Migration
 
             $this->createTable(Plugin::CACHE_TABLE, [
                 'uid'         => $this->string(40)->notNull()->unique(),
-                'urlHash'     => $this->string(32)->notNull(),
-                'url'         => $this->text()->notNull(),
+                'url'         => $this->string(255)->notNull(),
                 'headers'     => $this->text()->defaultValue(null),
                 'tags'        => $this->text()->notNull(),
                 'siteId'      => $this->integer(),
@@ -29,7 +28,7 @@ class Install extends Migration
                 'dateUpdated' => $this->dateTime()->null()
             ]);
 
-            $this->createIndex('url_idx', Plugin::CACHE_TABLE, 'urlHash', true);
+            $this->createIndex('url_idx', Plugin::CACHE_TABLE, 'url', true);
             $this->execute("ALTER TABLE " . Plugin::CACHE_TABLE . " ADD FULLTEXT INDEX tags_fulltext (tags ASC)");
 
         }
@@ -39,7 +38,6 @@ class Install extends Migration
 
             $this->createTable(Plugin::CACHE_TABLE, [
                 'uid'         => $this->string(40)->notNull()->unique(),
-                'urlHash'     => $this->string(32)->notNull(),
                 'url'         => $this->string(255)->notNull(),
                 'headers'     => $this->text()->defaultValue(null),
                 'tags'        => 'varchar[]',
@@ -49,7 +47,7 @@ class Install extends Migration
                 'PRIMARY KEY(uid)',
             ]);
 
-            $this->createIndex('url_idx', Plugin::CACHE_TABLE, 'urlHash', true);
+            $this->createIndex('url_idx', Plugin::CACHE_TABLE, 'url', true);
             $this->execute("CREATE INDEX tags_array ON " . Plugin::CACHE_TABLE . " USING GIN(tags)");
 
         }